Can build our deps with cargo. I think.
I did not update the rust submodule with this commit because I have a PR
waiting to be merged, after which we should be able to use vanilla Rust!

Deleted the unused crates, because their deps are liable to change by the
time we might use them in the future.